Parsing userEnrollmentStatus (#13)
Basically a bunch of changes to better accommodate the new `modelingPopulation` field - we had a field and were putting it in the temp table, but it was quietly failing because columns were in an unpredictable order. By manually specifying columns in the `INSERT` statements, we are more resilient to unpredictable changes in Twitter data. On a more optimistic notes, this will help make things continue working even if/when new data is added.

* log more details of errors to cloud

* log more details in all functions

* re-enable all functions

* create file to only test userEnrollmentStatus

* only use a very small number of rows for testing

* more accurate logs

* skip dropping table

* explicitly name columns and order for INSERT statement

* remove temp tables after inserting

* add column for modelPopulation if needed

* less aggressive dataframe dropping

* implement changes in main import script

* manually specify columns for ratings INSERT statement

* manually specify columns for notes INSERT

* test with a smaller number of rows

* manually specify columns when processing old data